Add CSVDataReader to read a csv file and refactor data reader
