+++ elasticdl/python/data/recordio_reader.py
@@ -1,0 +1,49 @@
+import os
+from contextlib import closing
+
+import recordio
+import tensorflow as tf
+
+from elasticdl.python.data.data_reader import (
+    AbstractDataReader,
+    Metadata,
+    check_required_kwargs,
+)
+
+
+class RecordIODataReader(AbstractDataReader):
+    def __init__(self, **kwargs):
+        AbstractDataReader.__init__(self, **kwargs)
+        self._kwargs = kwargs
+        check_required_kwargs(["data_dir"], self._kwargs)
+
+    def read_records(self, task):
+        with closing(
+            recordio.Scanner(
+                task.shard_name, task.start, task.end - task.start
+            )
+        ) as reader:
+            while True:
+                record = reader.record()
+                if record:
+                    yield record
+                else:
+                    break
+
+    def create_shards(self):
+        data_dir = self._kwargs["data_dir"]
+        start_ind = 0
+        f_records = {}
+        for f in os.listdir(data_dir):
+            p = os.path.join(data_dir, f)
+            with closing(recordio.Index(p)) as rio:
+                f_records[p] = (start_ind, rio.num_records())
+        return f_records
+
+    @property
+    def records_output_types(self):
+        return tf.string
+
+    @property
+    def metadata(self):
+        return Metadata(column_names=None)
